+++ CHANGELOG.md
@@ -16,6 +16,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.9.1 / 5.64.1] - 2023-04-??
 
+### Added
+- added option to disable all force rules of a given sandbox [#2797](https://github.com/sandboxie-plus/Sandboxie/issues/2797)
+
 ### Changed
 - reworked configuration storage for box grouping